+++ src/main/java/com/cloudapps/ecommerce/service/ShoppingCartService.java
@@ -32,7 +32,7 @@ public class ShoppingCartService {
 		return shoppingCartUseCase.findShoppingCartById(id);
 	}
 
-	public Optional<FullShoppingCartDto> addProduct(Long cartId, Long productId, Long prodQuantity) {
+	public Optional<FullShoppingCartDto> addProduct(Long cartId, Long productId, int prodQuantity) {
 		
 		return shoppingCartUseCase.addProduct(cartId, productId, prodQuantity);
 	}
